kanas: display: Change the lcd's physical size to a closer approximates
This is used to determine the dpi per dimensions which is used by
some apps for scaling

NOTE: The actuall dpi of this phone is about 207. You can go ahead
and change the ro.sf.lcd_density prop to 207 but be warned that the
screen elements will become smaller but would also give you more screen
real-estate. Could be a bad decision for the elderly or those with
bad eyesight , so I left it as it is.

Change-Id: Id2ce8408c3f66f53e3501534b1935a005a2a17b5
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>