kanas: patches: Bring s4k5ecgx mipi sensor to feature parity and HI702 ccir sensor.
Finally done it, at least the Enhanced Community Quarantine (EQC) here in the PH
gave me enough time to do research and fix up the sensor code to be on par
on the performance and feature with the stock vendor blob. Nevermind being
depressed for a week doing nothing to fix that wretched AF and blaming (read: cursing)
Samsung for it. (as it turns out, the AF was controlled by a separate regulator
and must to be turned on the right way. The fix? Just delete 34 lines.)
Sorry.

This patch set will also bring the nativehandle buffer type recording WHICH WILL
FIX the video recording on Android O once and for all. The credit goes to
djeman (github.com/djeman) as I searched his commits back in 2019 and found
something that actually could work but I did't dare to try since the s5k4ecgx
code is crippled.

Activate that fix using by setting that below on BoardConfig.mk
TARGET_USES_MEDIA_EXTENSIONS := true

On the other side HI702 should also have its WB settings, Image effects, and auto fps
that's all. Too few? The vendor blob side of this is also crippled anyways.
It should be possible to add exposure compensation to that sensor and fix the
resolutions presets according to my sources.

Change-Id: I51b94ae84ae60b0e2a52d31d5e59035add510762
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>