kanas: SPRDHardwareComposer: Blend up to 5 layers using GSP
Remember building the GSP hal patch before. This will be using it.
The SPRDevs intended to use the GSP with IOMMU since it works much better
than CMA (faster too) and it does not need contiguous memory. GSP
do need the layers to be contiguouly written but with the IOMMU.
IOMMU can transparently translate each access of the GSP as if it is
contiguous.

Unfortunately, we don't have that IOMMU goodness within kanas.
So, use a patch that will tell gralloc to allocate contiguous memory
for HWC layer buffers and enable some (disabled) patches like
that one called DEVICE_DIRECT_DISPLAY_SINGLE_OSD_LAYER which
borks the display on Android N.

There is some new patchset that reimplements the
sc8830 feature of blending multiple layers with GSP which is awesome.
It was remplemented since code-base as diverged so much.
The limit is currently set to 5.

NOTE: Requires a kernel patch to increase the CMA area.

Change-Id: Ic81823cb402d004600c6c031342c837746907f63
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>