kanas: healthd: rewrite the Charging slowly/rapidly messages patch
Due to some changes in healthd and Keyguard, the patch from Android N is
not needed anymore. Remember that the reason that patch was created is
that we always have a charging slowly message regardless when source
is used to charge the device.

But the patch (from Android N) has an extra feature of actually telling
that the phone is charging slowly if it's USB or AC. On Android O, the
changes will render the patch useless demanding a rewrite.
Even the overlay files have changed!

Change-Id: Id55f3bedf1cf2a263b600b25385736689e5cf0ae
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>