kanas: patches: redirect and build ril binaries from the scx15 folder
We don't even have modem_con...

Yeah, I'm considering switching over the opensource ril while
I'm still researching about the NVRAM switcheroo situation from
the stock rom.

Here a quick background:
The device have two partitions of the NVRAM: a main and a backup.
The Stock rom uses a different ril binary and uses the backup
as the main NVRAM. What's worse is that the ODIN firmware package,
writes the NVRAM on the backup rather than the main.

The opensource ril doesn't do any of that razzle-dazzle
so there may be cases where this ROM may actually bork the radio subsystem
where a reflash or restore of the NVRAM partitions may be needed.
It happened to be more than once on this ROM.
Both NVRAM Partition are wiped in my case, and have to be restored.
More testing needed.

Change-Id: If3bf39a2bc857df73ed26f2257139dacb5d523b9
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>