kanas: binderize composer and use the hwbinder for it and the allocator; Remove composer service
The reason the allocator fails under hwbinder transport is due to
SF_START_GRAPHICS_ALLOCATOR_SERVICE being unset.

Just verified and tested it, the allocator works with hwbinder as its transport.

The composer is still there, until we figure out why can't we use
the HWC2 mode (via the HWC2on1 adaptor) since the framebuffer_target
layer contains a NULL buffer handle as we can't always use the HWC for
everything.

The composer service, however, is removed since it starts parallel to
surfaceflinger and consumes memory that it will not use at all.
Surfaceflinger will only use the composer service if it's using HWC2
which we are not.

So for now, the Live Display support is still _TECHNICALLY_ broken.
It does work and browns down the screen but doing so will disable
SPRDHWComposer until reboot. Yes, even when you disable it.

Change-Id: If3985263a1c0ad191aa0e08853a7626b08c4f0a1
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>