kanas: ueventd: set character devices used by the camera to be actuall under the camera group
On the stock, it should have been enough setting the owner to system and
everything media related to audio. That assumption persists until
on Android O where more security measures are in place and the camera is handled
by cameraserver through android.hardware.camera.provider via the HIDL hwbinder transport.

To be safe and also to avoid greenscreens while using the camera,
set all the char devices that the camera HAL uses to be under the camera group.

Change-Id: I2e87b042035570fd611b1be34ec9e45a70bcb8c1
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>