kanas: Turn off TARGET_BOARD_CAMERA_DMA_COPY
It sounded nice that DMA is being used to copy redundant parts of memory
around for some reason but the sprd_dma_copy module does not copy
anything at all. Even after ensuring that it always gets a physicallycontiguous
src and dest buffers it still does not work. That explains the greenscreen
problem while using the camera.

Change-Id: I77651359f6fd488e52cf695453876ce3d91301ea
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>