kanas: Allow 1280x720p recording.
If we are using the opensource reimplementation of the s5k4ecx sensor
then it is possible to use the 720p recording without taxing the sprd_dcam
module of scaling up the 640x480 internal frame size since
the output frame size from the sensor is already 1280x720.

Plus, I forgot to build support for 960x720 scaling/recording in the
opensource camera hal.

Change-Id: I89f47f2c4471004240dfe3a74bb9956e3edf801b
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>