kanas: Allow SurfaceFlinger's Layer Triple Buffering
This is disabled by default.

I lost the logs but there are some apps that seem to use 3 buffers
which causes an error in setMaxAcquiredBufferCount().
Usually this occurs when the app's screen is replaced, destroyed or
multilated.

With this flag on (which increase the number of buffers), I can't find the
exact line anymore.

Change-Id: I07defc5477b736d0336a063a349ddd11c8fe5106
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>