kanas: Fix Low Power Mode (LPM) charging; few LPM tweaks
It's been bugging me for a while that at powered-off charging or LPM
I see the charger UI which is just a BW image of a battery charging.
That's what would happen if we're using the newer sprdfb kernel driver.
You WILL NOT get anything when using the _older_ that came with the stock
kernel since the charger UI seems to ab/use the blank() callback in the
driver which was not implemented in the older version.

Anyhow, because of some changes in the structure of Android O
the lpm service is now broken and has to be reappropiated to override
the previous charger UI and use the current structure (e.g. /system -> /vendor)

The few tweaks as metion in the title are:
	Use the powersave CPUFreq governor
	Disable the CPU hotplugging(*)
	Turn off all cores except for the boot CPU
        Tweak the SPRD's ondemand devfreq governor to make more inclined to choose
              the lowest DDR memory clock (200MHz)

(*) This assumes that the default hotplugging driver is the sprdemand cpufreq governor
    not some other hotplugging driver (like AUTOSMP, INTELLI-PLUG, etc...).
    Switching away from sprdemand governor will disable the hotplugger.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>