power: Workaround for the missing earlysuspend in libsuspend
Earlysuspend support was removed in libsuspend starting from Android O
leaving the wakeup_count as the only way to suspend the system.
...
Yeah, makes no sense to duplicate the effort I put in the comments of
the source code but the gist is that without the revert patch
from c94c8031bc197943fd7f74e1044d4611e756efc6, waking the device
will never work.

But I really wanted to lessen the changes we do to the platform code,
Like that patch bring the gralloc1/0 in libui which breaks the underlying
symlinks that Google dev/engg guys has created. In short, I'd like
less breakage due to directly editing their code.

Back to the earlysuspend issue, actually the kernel does support the wakeup_count
but its features while earlysuspend enabled is somewhat limited.
It can suspend the system but cannot wake it completely as it also needs
to activate the notifier_chain in earlysuspend's lateresume.
It can be triggered by simply writing "on" to '/sys/power/state'.

That's the crux of the WA; with this change, the patch to revert earlysuspend
should not be needed anymore. Additional goodies are also in place
along that workaround like detecting if we even need to to that WA
by locating the earlysuspend init symbol in libsuspend. It's not really
needed but hey, someone may decide to back/crossport this elsewhere.

Change-Id: I96d64561f01da1d50ba605ad6a994d050ec744f1
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>