bluetooth: Fix timed out commands
This partially reverts the 58f28ad88b1c35ec087bf76d1d7bddd1707c5194 as
the testing procedure is a bit skewed.

That commit actually breaks the BT system but gets fixed once the vendor
supplied libbt-vendor.so is used. At least, it does so until the next reboot.

This takes the default flags from Team Margarine's (or was it Team Butter's)
device tree for kanas as a base and also the old defaults from the Android 4.4
version.

Change-Id: Iae2ade4642fd1d38dfc60915a0a4ace980ab3f8b
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>