patches: Add a nifty way to fix the inverted frequency table used in newer encoders
TL;DR; It's possible to use the newer libomx_* encoders/decoders, well not the
really new ones (the one with VP9 encoding decoding). You just simply place
and use the newer code. But, the VSP frequency table is inverted with the highest
index sets the lowest and slowest clock and vice versa.

We could either patch the kernel for this or, just use a SHIM to intercept the
ioctl calls and make it "right", in a way.

Change-Id: I1c4c345070998f3360151fc0be55514f562a31ab
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>