kanas: gps: Fix agps crash and build hidl gnss services
I forgot that we already use the GNSS HIDL passthrough impl
and have not specified in hidl.mk that we need the services and the
corresponding init files. Without that, bootup will not be possible.

Regarding AGPS, the issue was already solved back on Marshmallow and
Nougat by forcing the platform code to use the AGpsInterface_v1
but on Android Oreo. They don't give a single ch*ck and go ahead
using the new-fangled data_conn_open_with_apn_ip_type() that does not
exists in AGpsInterface_v1.

The problem today is a bit less serious as this is a PURE null pointer
dereference since the platform code sees a 'cleaned' copy of the
AGpsInterface_v1 struct. Compared to before on Nougat (or maybe in Marshmallow),
Some actual code is being executed that is not data_conn_open_with_apn_ip_type()
which is quite dangerous given that we don't know exactly what is being executed.
A function at 0x0, just fails.

There is still a problem about Android not getting the list of XTRA server.
But that's for later.

Change-Id: Ib58f9b3b194b2656b2000c5f3f8faa9df528e5a9
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>