gps: Go back to using non binderized GNSS implementation
The 'gps' user that the GPS hal is running under is not allowed
to communicate to the network. Not even sockets are allowed to be created.

Even so, the GPS will still work as it is but AGPS will not as it will
fail to connect to the network all the time.

Revert back to the non-HIDL approach and let system_server handle the
GPS system.

Change-Id: Ib3e19693d10c61f766cacdf52a04d9dac3583f54
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>