Update TWRP fstab to include most backup-able partitions
This includes clean ups to the regular fstab and additions to the TWRP fstab.

Most of the additional partitions I've mentioned is used as a NVRAM
(Non Volatile Random Access Memory) for the coprocessor or the modem/radio
module. After experimenting with deferred resume and ignored in flight emmc
commands, I realized how important to have backups of these as well

As cautious as the vendor could be, aside from the NVRAM partitions
they also have backups or second partions to the main NVRAM partitions.
In total there are 4 main partitions: MODEM,DSP,RUNTIMENV,FIXNV; meaning
there are a total of 8 partitions. In a sense we're making
backups of the NVRAM backups too.

THe least honorable mention is that we will be able to backup the
/productinfo partition as well, which in my case got corrupted in my latest
venture of bringing up Android O.

P.S. A new version of TWRP is needed to install Android O and above which is
why this is being updated.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>