BoardConfig.mk: Automatically set kernel compilation flags
When kernel/samsung/kanas is present, we set the appropiate flags and
ignore the prebuilt kernel. But having designed (and reconstructed) this
tree with intents and purposes of solely building TWRP with the minimal
manifest.xml, I think the prebuilt kernel will suffice for most of the time
specially if the person's intent of building this is for his own use
and not for development.

Nevertheless, the kernel config used here is a specially crafted and minimal
TWRP_7.1_defconfig which is based on the original config by Samsung.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>