Add badblocks to the set of utilities in the recovery image
Rarely, the Android OS will start locking up for few seconds and this log will appear
on the kernel logs:

mmcblk0: error -110 transferring data, sector 1879872, nr 256, cmd response 0x900, card status 0xb00

To solve this issue you could just wipe the phone and reinstall the ROM or
follow the guide here.
https://forum.xda-developers.com/galaxy-s3/general/howto-fix-prevent-severe-freezes-wiping-t2566208

Though the author has provided a prebuilt version of badblocks, I'd say why not
include it in the recovery image from now on. And, this is where this patch comes in.

P.S. AFAIK, the Android O device tree I maintain has already included the badblocks executable.
If ever the system is still responsive, it's possible to perform the instructions from the
system without rebooting to recovery.

P.S. You can also use that method to just specify the specific sector to force-write.
But You'll need to calculate the block no. relative to the partition from the failing
sector. This is much more cost efficient in terms of write cycles but could be tedious
when there are many sectors indicated in the log.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>