kanas: Reduce LMK values
This was a patch from Android N which also has the same intention.
To reduce that 180MB line to 100MB. AFAIK, in Android KK, its about 128MB
then at some point in time, AOSP raised it up.

What's bothersome is that the boundary or reference point AOSP used in
calculating the buckets in LMK is at 700MB which made it
confusing if AOSP would take the high memory bucket or low memory one...

Change-Id: Id1472e4a2f7cf283f50ba8803b78c1c79f01dc6c
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>