kanas: patches: Revert support for extra heaps in gralloc
If it's just have a fallback, I should probably do the
libmemoryheapion_sprd library as well since some applications
go through that rather than gralloc. Plus, the fallback mechanism
was reimplemented in the kernel changing the userspace tools should not
be needed.

The Overlay memory being allocated in MM tweak was not removed, though.

Change-Id: I390e9459bd4b503d01557d76f48ee1a5a53a5a7d
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>