kanas: patches: Add missing scx35-common patches
These patchset contains various fixes. Two of them fixes the issues
of the current sepolicies ??????? (or maybe ????????'s younger sister)
has written or maintained. It should be possible to use enforcing mode
without breaking all of the functionality of the phone, only some of it.

The rest of the patches fixes the power hal memory leak and allows
people to disable it.

Change-Id: I43c37c66ba4a9450209d1c54ec4ec550b53738a2
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>