kanas: patches: remove the hardware_sprd patchset
It has been quite a burden trying to port and rebase these set of 70 patches
on the hardware_sprd tree (which hasn't been updated for about 2 years).
I think it will be more beneficial if we just fork out the repository and
apply and rebase the patches there.

And, in the Android O, the patches still kept growing and unfortunately
those patches are also applicable to Android P so yeah. having ~80 patches
are such a pain to maintain.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>