kanas: Initial commit for Android 9 (LineageOS 16.0)
Well not really. Android 8 and 9 branches are suppose to share the same
tree so we pretty much have already started since Android 7
(Android 5 and 6 was ih24n69's work anyways).

This commit will jump start the initial files to make the build system
"recognize" the kanas build target. Seriously, the Android 8 branch as
it is now have a hard time of building the LineageOS 16.0 source...
These few files are the first step.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>