kanas: Specify gralloc0 flag bits to ignore
There are some bits we use on the gralloc like the special flags to
indicate that the SPRD libcamera or the SPRD encoder/decoder
(e.g. SPRDAVCDecoder) or the SPRDHWComposer requests some special
buffers.

And.. there's also that flag that libMali.so always sets for no reason,
a flag went unused ages ago.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>