kanas: Create PGO profiles for this device
This also includes a new patch to allow overriding the default path at
toolchain/pgo-profiles which already contains all of the included or
pgo'd modules like libhwui, libskia, libhwbinder and dex2oat.

Also, these new profiles reflect the current state of AOSP P branch
on August 2020 rather than 2018 and is tested against the actual
device this tree is made for.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>