kanas: Try making GPS work by setting app_process SDK level to 19
The Marvell hybrid gps HAL driver relies on text reolcations and AOSP
is coded not to allow any of those stuff run and are blocked from the
linking stage. This is, IMHO, appropiate to discourage the use of
textrel... which should also improve security.

But our black box blob needs it.

As discussed in the comment in BoardConfig.mk, we must add an SDK version
override to the executable using that library blob.

For now try, adding an override to app_process, the executable that creates
zygote that in turn creates the system_server process.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>