kanas: rootdir: Mount /system 'on init' rather than 'on fs'
There's a bit of problem where some init *.rc files on /system or
/vendor are not being executed most notable those that starts on
late-init which happends before the fs trigger.

A notable example of this is atrace so you can pretty much expect that some
tracefs permissions are not set correctly.

This commit should modify the fstab a bit so mount_all --early and --late
can be used. On the init trigger, we'd try to mount the /system/ partition
as soon as possible. Then mount everything else on fs trigger as usual.

Note: This gives a possiblity of replacing that lost mount /system statement
on LPM mode.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>