kanas: usb: Implement a static USB and USB Gadget HAL
The default USB HAL 1.0 could work ?
With the default USB HAL doesn't really work with our device that don't
use USB-C (USB3 and stuff). We don't even have USB OTG, as far as I know.
Android won't, or doesn't seem to recognize, that a "configurable" usb
exists. What I meant about "configurable" is that USB plugin/plugout
will be detected but you may not switch modes like from charging to
MTP mode or even use USB thethering.

This USB hal presents a static configuration and mimics what Android expects
in newer phones (expect for the USB-C one). This allows us to set configurations
and stuff from Android without access to Developer Options.

Also, I'd also implement a rough C++ implementation of the some
sys.usb.state triggers in the init.sc8830.usb.rc init file. There are three
problems that I seek to solve:
    1. adbd gets reset whenever adb is enabled and we switch modes,
    mtp -> ptp or whatevs.
    2. mtp mode also activates acm function which throws off the legacy usb
    mode switcher in system_server. After this happens, you will not be
    able to set to any other modes through normal means.
    3. MTP function (in the kernel) does not work with FFS which adb uses.

The first one pertains to the stop adbd in init.usb.rc when sys.usb.config
is set to "none" which happens every time we switch modes. If you use
ADB wirelessly or through a network, this will temporarily break that.
It's annoying if you debug an app or collecting a trace then suddenly
forgot that you do and switch to MTP to copy some few songs in...

Second one is just how Samsung have modified the kernel code,
we could change it up but yeah. I'd prefer not to change it.

The third one is a bit heavy. When bringing up AOSP 9.0 or even 10.0 or 8.1
it makes things complicated when adb doesn't start at boot. Well, in
earlier versions, it works (5.1 and below) and seems to fails after
adb had switched to FFS (Function FS). You can test this out with this
USB Gadget HAL + USB HAL and switching over to MTP mode. ADB will be
unavailable despite having no error messages in the logs. Verify that any
other modes (PTP, tethering or even MIDI) will have adb but not MTP.

The solution? This commit will not be able to do that but switching over to
configFS will allow us to use both but it will require some patches to be
backported from 3.14.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>