kanas: gps: gps_wrapper: Convert to Android.bp
This is to match up the hidl implementation of gnss. That's it...

Well, excluding the fact that this is needed to debug some mysterious gps
problems because we use the hidl interface now...

Also, TARGET_SPECIFIC_HEADER_PATH doesn't work for Android.bp and
it's ticking both of them off as they are using the wrong headers.

And... I'm going to start calling commits to the gps wrapper code as
gps_wrapper.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: Ieaf6a50abd1f921e00bd7101753a9f203843165b