kanas: gps_wrapper: Do a runtime conversion on GpsSvStatus
Samsung has their own version that slips in about 4 bytes of memory
at the end of each GpsSvInfo. There are GPS_MAX_SVS GpsSvInfos in a
GpsSvStatus which ca draastically change how the Android GNSS system
interpret and show the data from the GPS.

We still get the location/GPS fix normally but we won't be able to read
the satelites it saw from any app. Honeslty, even Android will not be
able to parse it correctly.

So, do the conversions on runtime when ever the fixed gps.h header
ended up not being used.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I58e7318fa0751b17ceee534a20e37d042dc88eb4