kanas: gps_wrapper: Forcibly set the program's target sdk to 19 (KitKat)
Libraries and executables with text relocations are forbidden from
executing in an android device starting from Android M.

On Android M, N and O, LineageOS/CyanogenMod have a build time flag to
disable such check but they completely done it away starting Android 9
and opt to make it a whitelist in the linker executable.

We cannot whitelist this process since that whitelisting only applies to
executable. The library is being loaded directly and never used as
an executable. Unfortunately, the host process loading the library is
forked from app_process32 which means that the whitelist would also be
bypassed as it is only read once before the executable starts
Plus, the target sdk is reset to what ever sdk number is 10000 upon
forking.

The android_set_application_target_sdk_version() function is an
undocumented function that will forcibly change the target sdk and it's
used by the whitelisting I mentioned earlier. It's public and should be
usable elsewhere libdl is linked against meaning, we could also use that.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I5074f1fcba298b0a7d5ad8dc9484ab9d7513a70e