kanas: gps_wrapper: Compensate for the GPS week rollover
This is spotted by the GPSTest app (shoutout!) that the timestamps from
the GPS device is at February 2000 even though it's already September
2020.

That is because the GPS device will receive a week number from 0-1023
and at April 6 2019 (I guess) the number is incremented to 1024 but
because it only have 0-1023, the number is instead incremented back to 0.

Stil time cannot be moved (fortunately for us). I think it's safe to
assume a simple fix. Manually offset the time the GPS device reports
by 1024 weeks or 619315200000 milliseconds.

P.S. This patch will also kind of calibrate the time the GPS HAL reports.
It's always earlier than UTC. It's either my 1024 weeks offset is incorrect
or something.
GPS Time is usually AHEAD of UTC time and I'm not even sure what time
it reports if it's UTC or GPS time. This note assumes that its UTC.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I2da1b20a6c4e698f1d0a878a23c9ca5f287bf083