kanas: gps_wrapper: Fix AGpsRefLocation size mismatch
From the modified gps.h in scx35-common, AGpsRefLocation had an extra
member to indicate the Primary Scrambling Code (PSC)

Honestly, to support this, we will need to alter the JNI code, some
interfaces and the frameworks java code to allow setting this PSC.
This only way that I could think of is very intrusive so
just set this extra member to 2^16-1 or something.

This patch is the same with the patch that converts GpsSvStatus on runtime.
The conversion will only be done when the size of these structs differ.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I8e70a543580f95c337b832552eb789e094802098