kanas: Use FBIOGET_DMABUF and allow GSP to directly write to the framebuffer
These are really experimental flags with debatable benfit.

FBIOGET_DMABUF is an IOCTL that will export a dma-buf for mali to use
to 'directly' access the framebuffer without using any memcpy or so.
This may result in a performance boost but so far it doesn't seem to be
that noticable or I just simply suck at reliable testing.
The main reason to enable that is for us not to use the HIDL_INVALID_FD
workaround both in gralloc and SurfaceFlinger. The whole HIDL/binder
system can be used as it is intended. The flag is TARGET_HAS_FB_DMABUF.

On the other side, TARGET_GSP_ONTO_FB will allow HWC to write directly
to the Framebuffer which also brings questionable benefits.
Theoretically, it will bypass Spreadtrum's DISPC overlay feature
and allow GSP to use triple buffering. The first one could potentially
avoid the 'overhead' of laying out a physical buffer on screen
and timing it so that it will not tear. DISPC's overlay is
thought to be single buffered. The main benefit here is to use
less memory bandwidth to display a composite using GSP.
This is dependent on TARGET_HAS_FB_DMABUF.

NOTE: About DIRECT_DISPLAY_SINGLE_OSD feature in HWC at least when
TARGET_GSP_ONTO_FB=true, a copy to the framebuffer from a single
OSD layer will take hundreds of _micro_ seconds which looks like
a single DMA copy initiated by GSP. This is a good benefit, since
DIRECT_DISPLAY_SINGLE_OSD is riddled with two problems:
artifacts when the buffer is freed and tearing problems
in SPRD DISPC.

Since the buffer is copied onto the Framebuffer, no artifacts exists.
Since SPRD DISPC's overlay feature is not used, no tearing due to
triple buffering.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I5bcfa106b765ab68ba41e721f19f6b6806dc7fad