kanas: Initially allow adb to work without mtp
We still use the older non functionfs usb gadget framework in the kernel
meaning, we'd still rely on the f_mtp and the other f_* functions
when defining our usb composite device.

The problem here is that f_fs and f_mtp seem to be incompatible
that only f_mtp will work when f_fs is active. Note that we now use f_fs
instead of the f_adb for the adb functionality because Google has since
removed the support.

This issue will cause another problem when trying to use adb. Since,
the device has f_mtp overriding the f_fs, it will be impossible to use
adb until it is deactivated by switching USB modes in Android after
it booted. But, what if it hasn't or cannot boot properly.

As a workaround, we won't enable mtp by default. Well, by Android M, I think,
Android does not really enable MTP right away so this should be okay
to do.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I58aab8560c6375a6f21ab25c9e06c6bd4c38adb2