kanas: A disabled patch to add back the earlysuspend support in libsuspend
Yeap, "disabled". There are a couple of fixes mainly in the kernel to that
will allow earlysuspend and wakeup_count to work together without breaking
the suspend of the phone or any native apps that rely on earlysuspend.

Safe to say for now that the fixes are working without this patch being
applied. Though, I must say, the earlysuspend workaround in the powerhal
will be deactivated off once this patch is applied.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I978a11ec72688424a44d3c53317a5843cbed076f