kanas: Set TARGET_NEEDS_NETD_DIRECT_CONNECT_RULE
This is to maintain the previous behavior of netd setting up a default
route. I have no idea, yet again, if the modem needs this but based
on various experiences online, some RILs have a problem that
basically renders mobile data useless. Mobile data can be activated
and the link can be established but because the route is non
existent, the phone will not be able to connect to the internet.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I157c05e46661ed54c1a62d786dba43179509af65