kanas: Sync to working copy.
Testing for the systemui issue and vold, netd and logd cpu spinning issue
takes at least 3 full days (72+) and I still have no freakin' clue
why it happens or how it happens.

This commit is a snapshot and should bring the configuration up to the
same one used in the current test build.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I74104ff3c68122bbc06b9c7a9cb06303a75227a6
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>