kanas: Add a device_kanas_post.mk for post configuration
The way this is written, AOSP/Lineage configurations is built or set
upon our configuration (in the makefiles).
There should (or perhaps I could say instead "there will") be a time
where we may need to override or extend some of the flags used here.

Currently, the file just simply includes a comment.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: Icc6d45f1852b214b99b418b64b0953807f2455f5