kanas: Merge post config back to device.mk and order it last.
The reason why the device.mk is listed first is due to a bug that messes the
order of boot jars. It seems to be fixed since then so there's no reason to
make device.mk the first file to inherit its configuration.

Also telephony.mk is not any more inherited since it will be inherited
on common_full_phone.mk.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I4d42c54b8fdb799150617108d9f9e38c83f3f213