kanas: power: Try to run earlysuspend hook on set_interactive off
When using the workaround for the missing earlysuspend support in libsuspend,
the earlysuspend hooks may not be executed the moment the screen is turned off.
This is almost always true if the wakelocks implementation in the kernel
is based on wakeup_sources which will block reading to wakeup_count when
a wakeup_source is in flight.

It will eventually run the earlysuspend hooks when that happens but for the
mean time, the LCD, touchscreen and the such will remain powered on.

This patch will attempt to write "mem" to the /sys/power/state to run
the earlysuspend hooks as soon as the screen is turned off.
It should be safe to write that multiple times as the system
will only suspend when all wakeup_sources are "calm".

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I95ee9e5b62eb711e2a7b2f87b8bc5545b88c9e1f